--- front/src/App.css
@@ -1,12 +1,0 @@
-#root {
-    max-width: 1280px;
-    text-align: center;
-}
-
-.App {
-    position: relative;
-    display: flex;
-    flex-direction: row;
-    width: 100%;
-    height: 100%;
-}
