+++ front/src/pages/files/Files.tsx
@@ -1,0 +1,5 @@
+import style from '@/pages/files/files.module.css';
+
+export const Files = () => {
+    return <div>파일</div>;
+};
