+++ front/src/pages/workspace/workspace.module.css
@@ -1,0 +1,11 @@
+.default {
+    display: flex;
+    flex-direction: column;
+    width: 100vw;
+    height: 100vh;
+}
+
+.workspace {
+    display: flex;
+    flex-direction: row;
+}
